requirements.txt
Flask
numpy
darwin-py
opencv-python-headless